predicciones/requirements.txt
# requirements were added by pip freeze:
adium-theme-ubuntu==0.3.4
ansible==2.1.1.0
attrs==15.2.0
awscli==1.10.59
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
bleach==1.4.3
boto==2.43.0
boto3==1.4.1
botocore==1.4.70
cached-property==1.3.0
cairocffi==0.7.2
cffi==1.7.0
chardet==2.3.0
Chroma==0.2.0
colorama==0.3.7
coreapi==2.0.8
cryptography==1.5
cssutils==1.0.1
cycler==0.9.0
Cython==0.23.4
decorator==4.0.10
Django==1.10.3
django-appconf==1.0.2
django-bleach==0.3.0
django-braces==1.9.0
django-compressor==2.0
django-contrib-comments==1.7.1
django-cors-headers==1.1.0
django-cors-middleware==1.3.1
django-el-pagination==2.1.2
django-filter==0.15.3
django-guardian==1.4.4
django-haystack==2.5.2.dev32+ng201d27f
django-js-reverse==0.7.2
django-oauth-toolkit==0.10.0
django-push-notifications==1.4.1
django-rest-swagger==2.0.7
django-s3-folder-storage==0.5
django-statsd-mozilla==0.3.16
django-storages==1.5.1
django-tagging==0.4.3
django-widget-tweaks==1.4.1
django-wysiwyg-redactor==0.4.9.1
djangorestframework==3.5.3
docker-py==1.10.3
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
drf-haystack==1.6.0
drfdocs==0.0.11
ecdsa==0.13
edoo-odoo-integration==0.2.0
elasticsearch==2.4.0
enum34==1.1.6
et-xmlfile==1.0.1
funcsigs==0.4
functools32==3.2.3.post2
futures==3.0.5
geopy==1.11.0
gyp==0.1
html5lib==0.9999999
httpie==0.9.4
httplib2==0.9.2
idna==2.1
inlinestyler==0.2.1
ipaddress==1.0.17
ipython==4.2.0
ipython-genutils==0.1.0
itypes==1.1.0
jdcal==1.3
Jinja2==2.8
jmespath==0.9.0
jsonschema==2.5.1
lxml==3.6.0
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==1.5.1
mock==1.3.0
mutagen==1.31
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
netaddr==0.7.18
numpy==1.11.0
oauthlib==1.0.3
openapi-codec==1.1.7
openpyxl==2.4.0
PAM==0.4.2
paramiko==2.0.2
pathlib2==2.1.0
pbr==1.8.0
pep8==1.7.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pygame===1.9.1release
Pygments==2.1.3
pyOpenSSL==0.15.1
pyparsing==1.5.7
PyPDF2==1.26.0
pyserial==3.0.1
python-apt==1.1.0b1
python-dateutil==2.5.3
python-http-client==2.1.1
python-memcached==1.58
pytz==2016.6.1
PyYAML==3.12
rcssmin==1.0.6
reportlab==3.3.0
requests==2.11.0
rjsmin==1.0.12
route53==1.0.1
rsa==3.4.2
s3transfer==0.1.2
scipy==0.17.0
sendgrid==3.0.7
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.9.0
six==1.10.0
smtpapi==0.3.1
sorl-thumbnail==12.3
sqlparse==0.1.18
statsd==2.1.2
texttable==0.8.4
todobackend==0.1.0
traitlets==4.2.1
Twisted==16.0.0
txAMQP==0.6.2
Unidecode==0.4.19
unity-lens-photos==1.0
uritemplate==3.0.0
urllib3==1.18
uWSGI==2.0.14
validate-email==1.3
vboxapi==1.0
virtualenv==15.0.3
websocket-client==0.37.0
whisper==0.10.0rc1
xhtml2pdf==0.0.6
zope.interface==4.1.3

